Renamed RunTestBase to HydraBase
